refactor: move function
The `manifestPathFor` function not related to `UnityProjectManifests` themselves but to the IO of them. Moved into the corresponding module.
