+++ src/types/project-manifest.ts
@@ -1,11 +1,10 @@
-import { DomainName } from "./domain-name";
-import { SemanticVersion } from "./semantic-version";
-import { PackageUrl } from "./package-url";
-import { ScopedRegistry } from "./scoped-registry";
-import { RegistryUrl } from "./registry-url";
-import path from "path";
-import { removeTrailingSlash } from "../utils/string-utils";
-import { removeRecordKey } from "../utils/record-utils";
+import {DomainName} from "./domain-name";
+import {SemanticVersion} from "./semantic-version";
+import {PackageUrl} from "./package-url";
+import {ScopedRegistry} from "./scoped-registry";
+import {RegistryUrl} from "./registry-url";
+import {removeTrailingSlash} from "../utils/string-utils";
+import {removeRecordKey} from "../utils/record-utils";
 
 /**
  * The content of the project-manifest (manifest.json) of a Unity project.
@@ -155,15 +154,6 @@ export function addTestable(
     ...manifest,
     testables: [...(manifest.testables ?? []), name].sort(),
   };
-}
-
-/**
- * Determines the path to the package manifest based on the project
- * directory.
- * @param projectPath The root path of the Unity project.
- */
-export function manifestPathFor(projectPath: string): string {
-  return path.join(projectPath, "Packages/manifest.json");
 }
 
 /**
